requirements.txt
streamlit>=1.30
pdfplumber>=0.11
groq>=0.9
python-dotenv>=1.0